+++ yalc.lock
@@ -1,0 +1,9 @@
+{
+  "version": "v1",
+  "packages": {
+    "@esri/hub-common": {
+      "signature": "4a5e3a52d4a8e9032bce259034a05e90",
+      "file": true
+    }
+  }
+}
